Move K21 to tests
